(#154) Renomeando o arquivo de Autenticação da pasta service
